Move relationship normalizing into selectors
Also adds a basic selector to get all friends of the current user.
