+++ pages/api/relationships/index.ts
@@ -1,8 +1,8 @@
 import { prisma } from "@/db";
+import { normalizeRelationship } from "@/selectors/relationships";
 import { withUser } from "@/utility/apiAuth";
 import { User } from "@prisma/client";
 import { NextApiRequest, NextApiResponse } from "next";
-import { normalizeRelationship } from "./utility";
 
 export default withUser(async function handle(
   req: NextApiRequest,
@@ -30,7 +30,7 @@ export default withUser(async function handle(
       res.status(405).send({});
     }
   } catch (e) {
-      console.error(e);
-      return res.status(500).json({ success: false, error: e });
+    console.error(e);
+    return res.status(500).json({ success: false, error: e });
   }
 });
